requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
WTForms
Flask-Limiter
Flask-Talisman
python-dotenv
gunicorn
werkzeug
requests
itsdangerous
email-validator
psycopg2-binary
Flask-Mail
pandas
openpyxl
reportlab==3.6.13
